edit data model

+++ src/main/resources/migrations/changelog/V1__create_table_users.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users
 (
-    id INTEGER NOT NULL PRIMARY KEY,
+    id BIGINT NOT NULL PRIMARY KEY,
     input_data_state VARCHAR(30)
 );
